+++ client/src/components/css/Dashboard.css
@@ -35,8 +35,8 @@ border-radius: 1px solid #00916f;
     margin: 0 auto;
   }
 .part1{
-    width:80%;
-    height: 45%;
+    width:90%;
+    height: 55%;
     border-radius: 8px;
     border: 1px solid #00ba8e;
     margin: 2rem auto;
@@ -47,6 +47,13 @@ border-radius: 1px solid #00916f;
 .file{
     display: flex;
     flex-direction: column;
+}
+.upload-menu{
+    width: 90%;
+    margin: 1rem auto;
+}
+.upload-menu:focus{
+    outline: none;
 }
 .dashboard-file{
 width: 100%;
@@ -97,6 +104,7 @@ margin: 1rem auto;
 }
 .part2{
     display: block;
+    height: 45rem;
 }
 .collapse-button {
     display: inline;
@@ -143,7 +151,36 @@ margin: 1rem auto;
     align-items: center;
     justify-content: center;
 }
-
+.Notice-Download{
+    color: #00ba8e;
+}
+.Notice-Download:hover{
+    color: #00916f;
+}
+.read-less
+{
+display: none;
+margin: 1rem auto;
+}
+.read-more{
+display: block;
+}
+.showless
+{
+display: none;
+margin: 1rem auto;
+}
+.showmore{
+display: block;
+}
+.date{
+    position: relative;
+    /* right:-20rem */
+}
+.icon{
+    padding:1rem;
+    font-size: 1.5rem;
+}
 @media screen and (max-width: 1170px) {
     .Dashboard-profile{
         height:14rem;
@@ -185,7 +222,9 @@ margin: 1rem auto;
         width: 70%;
 
     }
-
+.Dashboard{
+    height: 50rem;
+}
     .profile-hr{
     display: none;
       }
@@ -199,8 +238,11 @@ margin: 1rem auto;
     }
     .part1{
         width:60%;
-        height: 75%;
+        height: 85%;
    
+    }
+    .part2{
+        height: 50rem;
     }
    
 }
